Moved Unused analyzers to Depreciated folder
